% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_internal_data.R
\docType{data}
\name{data_df}
\alias{data_df}
\title{data_df}
\format{
An object of class \code{data.frame} with 1453 rows and 11 columns.
}
\usage{
data_df
}
\description{
data_df
}
\keyword{datasets}
