% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_internal_data.R
\docType{data}
\name{new_data_df}
\alias{new_data_df}
\title{new_data_df}
\format{
An object of class \code{data.frame} with 1453 rows and 11 columns.
}
\usage{
new_data_df
}
\description{
new_data_df
}
\keyword{datasets}
